+++ generic_server.cc
@@ -21,6 +21,11 @@
 
 #include "generic_server.hh"
 
+#include "to_string.hh"
+
+#include <seastar/core/when_all.hh>
+#include <seastar/core/reactor.hh>
+
 namespace generic_server {
 
 connection::connection(server& server, connected_socket&& fd)
@@ -91,8 +96,9 @@ future<> connection::shutdown()
     return make_ready_future<>();
 }
 
-server::server(logging::logger& logger)
-    : _logger{logger}
+server::server(const sstring& server_name, logging::logger& logger)
+    : _server_name{server_name}
+    , _logger{logger}
 {
 }
 
@@ -120,6 +126,37 @@ future<> server::stop() {
         return on_stop();
     }).then([this] {
         return std::move(_stopped);
+    });
+}
+
+future<>
+server::listen(socket_address addr, std::shared_ptr<seastar::tls::credentials_builder> creds, bool is_shard_aware, bool keepalive) {
+    auto f = make_ready_future<shared_ptr<seastar::tls::server_credentials>>(nullptr);
+    if (creds) {
+        f = creds->build_reloadable_server_credentials([this](const std::unordered_set<sstring>& files, std::exception_ptr ep) {
+            if (ep) {
+                _logger.warn("Exception loading {}: {}", files, ep);
+            } else {
+                _logger.info("Reloaded {}", files);
+            }
+        });
+    }
+    return f.then([this, addr, is_shard_aware, keepalive](shared_ptr<seastar::tls::server_credentials> creds) {
+        listen_options lo;
+        lo.reuse_address = true;
+        if (is_shard_aware) {
+            lo.lba = server_socket::load_balancing_algorithm::port;
+        }
+        server_socket ss;
+        try {
+            ss = creds
+                ? seastar::tls::listen(std::move(creds), addr, lo)
+                : seastar::listen(addr, lo);
+        } catch (...) {
+            throw std::runtime_error(format("{} error while listening on {} -> {}", _server_name, addr, std::current_exception()));
+        }
+        _listeners.emplace_back(std::move(ss));
+        _stopped = when_all(std::move(_stopped), do_accepts(_listeners.size() - 1, keepalive, addr)).discard_result();
     });
 }
